Edits.

+++ Defender/src/CollisionDetector/CollisionDetector.java
@@ -5,6 +5,7 @@ import GameObjects.*;
 import javafx.scene.shape.Rectangle;
 import java.util.ArrayList;
 
+
 public class CollisionDetector {
     private static CollisionDetector collisionDetector = null;
 
@@ -28,17 +29,18 @@ public class CollisionDetector {
         Rectangle motherBox = motherShip.getHitbox();
 
         for(Projectile projectile: projectiles){
-            Rectangle projectileBox = projectile.getHitbox();
-            if(motherShip.isAlive() && projectile.isAlive() && projectileBox.getBoundsInParent().intersects(motherBox.getBoundsInParent())){
-                int damage = projectile.getDamage();
-                motherShip.updateHealth(damage);
+            if ( !(projectile instanceof ShipProjectile) ) {
+                Rectangle projectileBox = projectile.getHitbox();
+                if (motherShip.isAlive() && projectile.isAlive() && projectileBox.getBoundsInParent().intersects(motherBox.getBoundsInParent())) {
+                    int damage = projectile.getDamage();
+                    motherShip.updateHealth(damage);
 
-                if(motherShip.getHealth() <= 0){
-                    motherShip.kill();
-                    break;
+                    if (motherShip.getHealth() <= 0) {
+                        motherShip.kill();
+                        break;
+                    }
+                    projectile.kill();
                 }
-                projectile.kill();
-                break;
             }
         }
     }
@@ -63,9 +65,11 @@ public class CollisionDetector {
             Rectangle humanBox = human.getHitbox();
             for(Alien alien: aliens){
                 Rectangle alienBox = alien.getHitbox();
+
                 if(alien.isAlive() && human.isAlive() && alienBox.getBoundsInParent().intersects(humanBox.getBoundsInParent())){
                     alien.kill();
                     human.kill();
+                    // over for this alien;
                     break;
                 }
             }
